refactor(renderer): MessageItem 拆分 message-actions/UserMessage/SystemMessage，readOnly/busy selector 收敛为 hooks
